add free variables to conditional structures
