+++ src/test/resources/resolver/xmirs/conditional_attribute/cond.xmir
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<program ms="393"
+<program ms="245"
           name="transformer"
-          time="2022-09-25T14:39:12.102246610Z"
+          time="2022-09-26T14:35:50.953996513Z"
           version="0.28.5">
    <listing>[] &gt; base
   [] &gt; y
